process_service/requirements.txt
Package               Version
--------------------- -----------
amqp                  5.3.1
annotated-types       0.7.0
ansicon               1.89.0
appdirs               1.4.4
asgiref               3.11.1
async-timeout         5.0.1
attrs                 21.2.0
beautifulsoup4        4.11.1
billiard              4.2.4
blessed               1.19.0
blinker               1.9.0
branca                0.7.0
cattrs                1.8.0
celery                5.6.2
certifi               2021.5.30
charset-normalizer    2.0.7
click                 8.1.8
click-didyoumean      0.3.1
click-plugins         1.1.1.2
click-repl            0.3.0
colorama              0.4.4
commonmark            0.9.1
confluent-kafka       2.13.2
distlib               0.3.2
Django                4.2.29
django-ninja          1.5.1
djangorestframework   3.16.1
docutils              0.22.4
exceptiongroup        1.3.1
filelock              3.0.12
Flask                 3.1.0
flit                  3.12.0
flit_core             3.12.0
flyingcircus          0.1.4.0
folium                0.15.0
idna                  3.3
importlib_metadata    8.6.1
itsdangerous          2.2.0
Jinja2                3.1.2
jinxed                1.1.0
kombu                 5.6.2
MarkupSafe            2.1.3
MouseInfo             0.1.3
mypy                  1.10.1
mypy-extensions       1.0.0
numpy                 1.21.4
opencv-python         4.5.4.60
packaging             21.0
pandas                1.4.3
pika                  1.3.2
pillow                11.1.0
pip                   26.0.1
pipenv                2021.5.29
prettytable           3.0.0
prompt_toolkit        3.0.52
psycopg2              2.9.1
psycopg2-binary       2.9.11
PyAutoGUI             0.9.54
pydantic              2.12.5
pydantic_core         2.41.5
PyGetWindow           0.0.9
Pygments              2.11.2
PyMsgBox              1.0.9
pyparsing             2.4.7
pyperclip             1.9.0
pyproject_hooks       1.2.0
PyRect                0.2.0
PyScreeze             1.0.1
python-abc            0.2.0
python-dateutil       2.8.2
python-dotenv         1.0.1
pytk                  0.0.2.1
pytweening            1.2.0
pytz                  2021.1
pywhatkit             5.4
pywin32               303
redis                 7.0.1
requests              2.26.0
requests-cache        0.8.1
rich                  11.1.0
setuptools            56.0.0
setuptools-scm        6.3.2
six                   1.16.0
soupsieve             2.3.2.post1
sqlparse              0.4.1
tomli                 1.2.1
tomli_w               1.2.0
typing_extensions     4.15.0
typing-inspection     0.4.2
tzdata                2025.3
tzlocal               5.3.1
url-normalize         1.4.3
urllib3               1.26.7
uv                    0.9.3
vine                  5.1.0
virtualenv            20.4.7
virtualenv-clone      0.5.4
virtualenvwrapper-win 1.2.7
wcwidth               0.2.5
Werkzeug              3.1.3
wikipedia             1.4.0
zipp                  3.21.0
